test TestInstructionTypeFromOpcode
